reorg _spaces namespace
