+++ src/orquestra/sdk/_base/_api/__init__.py
@@ -10,14 +10,12 @@ We re-export symbols here for grouping concepts under the "api" umbrella, e.g.
 
 from ._config import RuntimeConfig, migrate_config_file
 from ._task_run import TaskRun
-from ._wf_run import WorkflowRun, list_projects, list_workflow_runs, list_workspaces
+from ._wf_run import WorkflowRun, list_workflow_runs
 
 __all__ = [
     "RuntimeConfig",
     "TaskRun",
     "WorkflowRun",
     "list_workflow_runs",
-    "list_projects",
-    "list_workspaces",
     "migrate_config_file",
 ]
