+++ src/orquestra/sdk/_base/_spaces/__init__.py
@@ -1,0 +1,14 @@
+################################################################################
+# © Copyright 2023 Zapata Computing Inc.
+################################################################################
+
+from ._api import list_projects, list_workspaces
+from ._structs import Project, ProjectRef, Workspace
+
+__all__ = [
+    "list_workspaces",
+    "list_projects",
+    "ProjectRef",
+    "Project",
+    "Workspace",
+]
